Remove unused custom HTML attribute
chrSetIndex is currently merely chrIndex.  It might be needed later (e.g. to enable rotation in polyploid ideograms), but currently it is not used.  And if chrSetIndex is ultimately used, a custom HTML attribute is likely not the best place to store that data.  So remove HTML attribute, but preserve argument in various function signatures.